change files to ts to increase type validation
