+++ src/server/routes/entity/series.ts
@@ -30,7 +30,7 @@ import {
 } from '../../helpers/entityRouteUtils';
 
 import {ConflictError} from '../../../common/helpers/error';
-import {PrivilegeTypes} from '../../../common/helpers/privileges-utils';
+import {PrivilegeType} from '../../../common/helpers/privileges-utils';
 import _ from 'lodash';
 import {escapeProps} from '../../helpers/props';
 import express from 'express';
@@ -40,6 +40,12 @@ import target from '../../templates/target';
 /** ****************************
 *********** Helpers ************
 *******************************/
+
+type PassportRequest = express.Request & {
+	user: any,
+	session: any
+};
+
 const additionalSeriesProps = [
 	'entityType', 'orderingTypeId'
 ];
@@ -80,7 +86,7 @@ const mergeHandler = makeEntityCreateOrEditHandler(
 	'series', transformNewForm, additionalSeriesProps, true
 );
 
-const ENTITY_EDITOR = PrivilegeTypes.ENTITY_EDITING_PRIV.value;
+const {ENTITY_EDITOR} = PrivilegeType;
 
 /** ****************************
 *********** Routes ************
@@ -93,7 +99,7 @@ router.get(
 	'/create', auth.isAuthenticated, auth.isAuthorized(ENTITY_EDITOR), middleware.loadIdentifierTypes,
 	middleware.loadLanguages,
 	middleware.loadRelationshipTypes, middleware.loadSeriesOrderingTypes,
-	async (req, res) => {
+	async (req: PassportRequest, res) => {
 		const {markup, props} = entityEditorMarkup(generateEntityProps(
 			'series', req, res, {}
 		));
@@ -130,7 +136,7 @@ router.get(
 router.post(
 	'/create', entityRoutes.displayPreview, auth.isAuthenticatedForHandler, auth.isAuthorized(ENTITY_EDITOR), middleware.loadIdentifierTypes,
 	middleware.loadLanguages,
-	middleware.loadRelationshipTypes, middleware.loadSeriesOrderingTypes, async (req, res) => {
+	middleware.loadRelationshipTypes, middleware.loadSeriesOrderingTypes, async (req: PassportRequest, res) => {
 		const entity = await utils.parseInitialState(req, 'series');
 		if (entity.seriesSection) {
 			const orderingTypes = ['Automatic', 'Manual'];
@@ -200,12 +206,12 @@ function _setSeriesTitle(res) {
 }
 
 router.get('/:bbid', middleware.loadEntityRelationships, middleware.loadSeriesItems, middleware.loadGenders,
-	middleware.loadWikipediaExtract, (req, res) => {
+	middleware.loadWikipediaExtract, (req: PassportRequest, res) => {
 		_setSeriesTitle(res);
 		entityRoutes.displayEntity(req, res);
 	});
 
-router.get('/:bbid/delete', auth.isAuthenticated, auth.isAuthorized(ENTITY_EDITOR), (req, res, next) => {
+router.get('/:bbid/delete', auth.isAuthenticated, auth.isAuthorized(ENTITY_EDITOR), (req: PassportRequest, res, next) => {
 	if (!res.locals.entity.dataId) {
 		return next(new ConflictError('This entity has already been deleted'));
 	}
@@ -215,7 +221,7 @@ router.get('/:bbid/delete', auth.isAuthenticated, auth.isAuthorized(ENTITY_EDITO
 
 router.post(
 	'/:bbid/delete/handler', auth.isAuthenticatedForHandler, auth.isAuthorized(ENTITY_EDITOR),
-	(req, res) => {
+	(req: PassportRequest, res) => {
 		const {orm} = req.app.locals;
 		const {SeriesHeader, SeriesRevision} = orm;
 		return entityRoutes.handleDelete(
@@ -224,13 +230,13 @@ router.post(
 	}
 );
 
-router.get('/:bbid/revisions', (req, res, next) => {
+router.get('/:bbid/revisions', (req: PassportRequest, res, next) => {
 	const {SeriesRevision} = req.app.locals.orm;
 	_setSeriesTitle(res);
 	entityRoutes.displayRevisions(req, res, next, SeriesRevision);
 });
 
-router.get('/:bbid/revisions/revisions', (req, res, next) => {
+router.get('/:bbid/revisions/revisions', (req: PassportRequest, res, next) => {
 	const {SeriesRevision} = req.app.locals.orm;
 	_setSeriesTitle(res);
 	entityRoutes.updateDisplayedRevisions(req, res, next, SeriesRevision);
@@ -304,7 +310,7 @@ export function seriesToFormState(series) {
 		}
 	});
 
-	const optionalSections = {};
+	const optionalSections: any = {};
 	if (series.annotation) {
 		optionalSections.annotationSection = series.annotation;
 	}
@@ -324,7 +330,7 @@ router.get(
 	'/:bbid/edit', auth.isAuthenticated, auth.isAuthorized(ENTITY_EDITOR), middleware.loadIdentifierTypes,
 	middleware.loadSeriesOrderingTypes, middleware.loadLanguages,
 	 middleware.loadEntityRelationships, middleware.loadRelationshipTypes,
-	(req, res) => {
+	(req: PassportRequest, res) => {
 		const {markup, props} = entityEditorMarkup(generateEntityProps(
 			'series', req, res, {}, seriesToFormState
 		));
