+++ src/server/routes/unifiedform.ts
@@ -1,13 +1,13 @@
 import * as middleware from '../helpers/middleware';
 import {createEntitiesHandler, generateUnifiedProps, unifiedFormMarkup} from '../helpers/entityRouteUtils';
 import {isAuthenticated, isAuthenticatedForHandler, isAuthorized} from '../helpers/auth';
-import {PrivilegeTypes} from '../../common/helpers/privileges-utils';
+import {PrivilegeType} from '../../common/helpers/privileges-utils';
 import {escapeProps} from '../helpers/props';
 import express from 'express';
 import target from '../templates/target';
 
 
-const ENTITY_EDITOR = PrivilegeTypes.ENTITY_EDITING_PRIV.value;
+const {ENTITY_EDITOR} = PrivilegeType;
 
 type PassportRequest = express.Request & {user: any, session: any};
